# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage PDF documents",
        "description": "Simple API for manage PDF files",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/types/": {
            "get": {
                "summary": "List types",
                "description": "Returns a list of all document types registered",
                "operationId": "list_types",
                "responses": {
                    "200": {
                        "description": "Showing all document types"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/types": {
            "post": {
                "summary": "create a document types",
                "description": "create a document types",
                "operationId": "new_Type",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "document types created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        },
        "/api/types/{id}": {
            "get": {
                "summary": "Get document type",
                "description": "Returns information about document type by id",
                "operationId": "get_Type_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "document types retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            },
            "put": {
                "summary": "update a document types",
                "description": "update a document types",
                "operationId": "update_Type",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "document types updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            },
            "delete": {
                "summary": "delete a document types",
                "description": "delete a document types",
                "operationId": "delete_Type",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "document type deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server error, try again"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TypeRequest": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "string",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
...
            "TypeSchema": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "string",
                        "nullable": "false"
                    }
                },
                "type": "object"
            }
        }
    }
}